# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/wallets/bank-account": {
      "put": {
        "operationId": "WalletsController_updateBankAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBankAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bank account updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update bank account information",
        "tags": [
          "Wallets"
        ]
      },
      "get": {
        "operationId": "WalletsController_getBankAccount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bank account details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get masked bank account",
        "tags": [
          "Wallets"
        ]
      }
    },
    "/wallets/balance": {
      "get": {
        "operationId": "WalletsController_getBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Wallet balance.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current wallet balance",
        "tags": [
          "Wallets"
        ]
      }
    },
    "/wallets/deposit": {
      "post": {
        "operationId": "WalletsController_deposit",
        "parameters": [
          {
            "name": "x-challenge-token",
            "in": "header",
            "description": "Challenge Token obtained from /auth/verify-otp",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositWithdrawDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deposit transaction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deposit Fiat via Mock Stripe",
        "tags": [
          "Wallets"
        ]
      }
    },
    "/wallets/withdraw": {
      "post": {
        "operationId": "WalletsController_withdraw",
        "parameters": [
          {
            "name": "x-challenge-token",
            "in": "header",
            "description": "Challenge Token obtained from /auth/verify-otp",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositWithdrawDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Withdraw transaction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Withdraw Fiat via Mock Stripe",
        "tags": [
          "Wallets"
        ]
      }
    },
    "/wallets/transactions": {
      "get": {
        "operationId": "WalletsController_getTransactions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTransactionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transaction history",
        "tags": [
          "Wallets"
        ]
      }
    },
...
    "/bounty/lock-fund": {
      "post": {
        "operationId": "Web3Controller_lockFund",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockFundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Funds locked successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Web3ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lock funds in Smart Contract (Escrow)",
        "tags": [
          "Web3 & Bounty"
        ]
      }
    },
    "/bounty/approve-payout": {
      "post": {
        "operationId": "Web3Controller_approvePayout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovePayoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payout approved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Web3ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve payout from Smart Contract to Developer",
        "tags": [
          "Web3 & Bounty"
        ]
      }
    },
    "/bounty/sync": {
      "post": {
        "operationId": "Web3Controller_syncTransaction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction synced.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "summary": "Sync Blockchain Transaction Status (Webhook/Indexer)",
        "tags": [
          "Web3 & Bounty"
        ]
      }
    },
...
    "/webhooks/github": {
      "post": {
        "operationId": "WebhooksController_handleGithubWebhook",
        "parameters": [
          {
            "name": "x-github-event",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-hub-signature-256",
            "in": "header",
            "description": "GitHub HMAC signature",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Webhook received and processed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDto"
                }
              }
            }
          }
        },
        "summary": "Handle GitHub Webhooks (e.g. push events)",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/webhooks/gitlab": {
      "post": {
        "operationId": "WebhooksController_handleGitlabWebhook",
        "parameters": [
          {
            "name": "x-gitlab-event",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-gitlab-token",
            "in": "header",
            "description": "GitLab Secret Token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Webhook received and processed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponseDto"
                }
              }
            }
          }
        },
        "summary": "Handle GitLab Webhooks (e.g. pipeline events)",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Bounty Platform API",
    "description": "The Nền tảng Quản lý công việc và Giao việc tự do API description.\n\n### Real-time Events (WebSocket)\nThis platform provides real-time updates via Socket.io.\n- **Endpoint**: `ws://<host>:<port>/notifications`\n- **Transports**: `['websocket', 'polling']`\n\n#### Events to Listen For\n1. **`risk-alert`**: Emitted when the AI detects a high risk score for a task.\n   - Payload: `{ pmId: string, message: string }`\n2. **`balance-warning`**: Emitted when the PM's escrow balance drops below a required threshold.\n   - Payload: `{ pmId: string, message: string }`\n",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateBankAccountDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "example": "123456789"
          },
          "bankName": {
            "type": "string",
            "example": "Techcombank"
          }
        },
        "required": [
          "accountNumber",
          "bankName"
        ]
      },
      "BankAccountResponseDto": {
        "type": "object",
        "properties": {
          "maskedData": {
            "type": "string",
            "example": "**** 6789"
          }
        },
        "required": [
          "maskedData"
        ]
      },
      "BalanceResponseDto": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "number",
            "example": 1050.5
          }
        },
        "required": [
          "balance"
        ]
      },
      "DepositWithdrawDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 500
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "nonce": {
            "type": "string",
            "example": "nonce-uuid-1234"
          },
          "bankAccountId": {
            "type": "string",
            "example": "bank-account-uuid"
          }
        },
        "required": [
          "amount",
          "currency",
          "nonce"
        ]
      },
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-tx-1234"
          },
          "userId": {
            "type": "string",
            "example": "uuid-user-1234"
          },
          "type": {
            "type": "string",
            "example": "DEPOSIT"
          },
          "amount": {
            "type": "number",
            "example": 500
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "status": {
            "type": "string",
            "example": "PENDING"
          },
          "txHash": {
            "type": "string",
            "example": "0xhash..."
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "amount",
          "currency",
          "status",
          "txHash"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "lastPage": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "lastPage"
        ]
      },
      "PaginatedTransactionResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "LockFundDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "example": "task-uuid-123"
          }
        },
        "required": [
          "taskId"
        ]
      },
      "Web3ResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "txHash": {
            "type": "string",
            "example": "0xlock_123456789"
          },
          "message": {
            "type": "string",
            "example": "Funds locked in Escrow"
          }
        },
        "required": [
          "success",
          "txHash",
          "message"
        ]
      },
      "ApprovePayoutDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "example": "task-uuid-123"
          }
        },
        "required": [
          "taskId"
        ]
      },
      "SyncTransactionDto": {
        "type": "object",
        "properties": {
          "txHash": {
            "type": "string",
            "example": "0x123abc456def789..."
          },
          "status": {
            "type": "string",
            "example": "COMPLETED",
            "enum": [
              "COMPLETED",
              "FAILED"
            ]
          }
        },
        "required": [
          "txHash",
          "status"
        ]
      },
      "SyncResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Transaction status updated to COMPLETED"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "WebhookResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      }
    }
  }
}